ImportKeys: Extract interface and refactoring
ImportKeys: Extract interface and refactoring
